create a script loader for bootstrap
all scricpts from `boot.d` will be executed once, and only once, as
a file will be stored in `$HOME/.boot_log` to keep track what have
already been executed. So if one want to run a script again, one
will have to delete the script entry in `.boot_log`.

wip:
in order to know if a script is already executed a method quite
similar to what have been done in brew.sh is used (array and a
method to check if the element is in the array), so it might be good
to mutualize all of that at some point (same for logs, and bold,
normal variables).