adds NODE_PATH env var
this is required to know where the global node_modules directory is
(for example `fx` need this).

So old `node.env` which is responsible to add the `.bin` directory of
the local node_modules in the path, have been renamed into
`node_bin.env`. So from now on `node.env` is really some command to
tune node.

wip:
- need to understand why nvp is not recognize in scripts, the
  current nvm had to be hardcoded, and this is a pity as it was
supposed to be dynamically loaded using `$(nvm current)`.