the passphrase for ssh key is asked everytime... try to remember it
something changed some months ago, from now on this is asking the
passphrase way too often.
By running the agent, it seems to ask it only one per terminal
session. Would be great to ask it once and not one per terminal
session ... but anyway this is better than nothing.