auto rearrange spaces when switching from 1 to 3 displays
if 1 display, left monitor windows go to space 3, and right monitor
windows to space 2.
When switching to 3 displays, we do the opposite.

And we are storing the state every time a window is modified using
yabay signals.

wip: it would be nice to also manage a 2 display setting. even if
this is a rare setting in my daily routine.