improve yabai restoring layout
using two files, one per number of displays was not a good idea,
because when exiting sleep, we would try to restore for 3 displays
and analyze the 1 display dump. Which might have been staled (as I
work 90% of the time with the 3 displays). So every time I was
exiting sleep, if I was having a wrong staled 1 display dump, windows
were move in wrong places.

Instead of doing this, we are now using only one dump and
translating on save and restore based on the number of displays. The
dump always use the max number of display, so 3.

Additionally some logs have been created when running the script by
yabai triggers.