fix: display position
adds delay before dumping positions. The dump should happen after a
transform if we exit sleep.
With the new single file, we were having some issues when we were
exiting sleep, everything was set on main display, so it was
triggering some dump and when the transform was applied it was
taking the latest dump into account and not putting back windows on
external displays.
With the delay, it seems to be way better...

As sometimes we still have the issue, a new parameter to the script
has been added.
```
window_reorg fix
```
It will force all windows in space 2 to go to space 11 and all from
3 to space 12.